fix(otelgenai): emit gen_ai.request.top_k as integer (#680)
Fix `gen_ai.request.top_k` type, it should be an
[integer](https://github.com/open-telemetry/semantic-conventions-genai/blob/main/docs/registry/attributes/gen-ai.md#gen-ai-request-top-k)
([changed](https://github.com/open-telemetry/semantic-conventions-genai/pull/217)).

+++ go/otelgenai/event_test.go
@@ -198,7 +198,7 @@ func TestOperationDetailsEventStructuredContent(t *testing.T) {
 		t.Error("event context has no valid span context")
 	}
 	attrs := logAttributes(record)
-	if got := attrs["gen_ai.request.top_k"].AsFloat64(); got != 40 {
+	if got := attrs["gen_ai.request.top_k"].AsInt64(); got != 40 {
 		t.Errorf("gen_ai.request.top_k = %v, want 40", got)
 	}
